+++ app/upgrades/v24/upgrades.go
@@ -2,7 +2,6 @@ package v24
 
 import (
 	"sort"
-	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
@@ -11,8 +10,8 @@ import (
 
 	"github.com/osmosis-labs/osmosis/v24/app/keepers"
 	"github.com/osmosis-labs/osmosis/v24/app/upgrades"
-	v23 "github.com/osmosis-labs/osmosis/v24/app/upgrades/v23"
 	concentratedliquidity "github.com/osmosis-labs/osmosis/v24/x/concentrated-liquidity"
+	concentratedtypes "github.com/osmosis-labs/osmosis/v24/x/concentrated-liquidity/types"
 	cwpooltypes "github.com/osmosis-labs/osmosis/v24/x/cosmwasmpool/types"
 	incentivestypes "github.com/osmosis-labs/osmosis/v24/x/incentives/types"
 	txfeestypes "github.com/osmosis-labs/osmosis/v24/x/txfees/types"
@@ -56,23 +55,11 @@ func CreateUpgradeHandler(
 		// and spread the pruning time across multiple blocks to avoid a single block taking too long.
 		keepers.TwapKeeper.SetDeprecatedHistoricalTWAPsIsPruning(ctx)
 
-		// restrict lockable durations to 2 weeks per
-		// https://wallet.keplr.app/chains/osmosis/proposals/400
-		chainID := ctx.ChainID()
-
-		if chainID == mainnetChainID || chainID == edgenetChainID {
-			keepers.IncentivesKeeper.SetLockableDurations(ctx, []time.Duration{
-				time.Hour * 24 * 14,
-			})
-			keepers.PoolIncentivesKeeper.SetLockableDurations(ctx, []time.Duration{
-				time.Hour * 24 * 14,
-			})
-		}
-
 		// Set the new min value for distribution for the incentives module.
 		// https://www.mintscan.io/osmosis/proposals/733
 		keepers.IncentivesKeeper.SetParam(ctx, incentivestypes.KeyMinValueForDistr, incentivestypes.DefaultMinValueForDistr)
 
+		chainID := ctx.ChainID()
 		// We only perform the migration on mainnet pools since we hard-coded the pool IDs to migrate
 		// in the types package. And the testnet was migrated in v24
 		if chainID == mainnetChainID || chainID == edgenetChainID {
@@ -85,8 +72,8 @@ func CreateUpgradeHandler(
 
 		// White Whale uploaded a broken contract. They later migrated cwpool via the governance
 		// proposal in x/cosmwasmpool
-		// However, there was a problem in the migration logic where the CosmWasmpool state CodeId  did not get updated.
-		// As a result, the CodeID for the contract that is tracked in x/wasmd  was migrated correctly. However, the code ID that we track in the x/cosmwasmpool  state did not.
+		// However, there was a problem in the migration logic where the CosmWasmpool state CodeId did not get updated.
+		// As a result, the CodeID for the contract that is tracked in x/wasmd was migrated correctly. However, the code ID that we track in the x/cosmwasmpool state did not.
 		// Therefore, we should perform a migration for each of the hardcoded white whale pools.
 		poolIds := []uint64{1463, 1462, 1461}
 		for _, poolId := range poolIds {
@@ -120,8 +107,8 @@ func CreateUpgradeHandler(
 
 // migrateMainnetPools migrates the specified mainnet pools to the new accumulator scaling factor.
 func migrateMainnetPools(ctx sdk.Context, concentratedKeeper concentratedliquidity.Keeper) error {
-	poolIDsToMigrate := make([]uint64, 0, len(FinalIncentiveAccumulatorPoolIDsToMigrate))
-	for poolID := range FinalIncentiveAccumulatorPoolIDsToMigrate {
+	poolIDsToMigrate := make([]uint64, 0, len(concentratedtypes.MigratedIncentiveAccumulatorPoolIDsV24))
+	for poolID := range concentratedtypes.MigratedIncentiveAccumulatorPoolIDsV24 {
 		poolIDsToMigrate = append(poolIDsToMigrate, poolID)
 	}
 
@@ -143,7 +130,7 @@ func migrateMainnetPools(ctx sdk.Context, concentratedKeeper concentratedliquidi
 		}
 
 		// This should never happen, this check is defence in depth in case we have wrong data by accident
-		_, isMigrated := v23.MigratedIncentiveAccumulatorPoolIDs[poolID]
+		_, isMigrated := concentratedtypes.MigratedIncentiveAccumulatorPoolIDs[poolID]
 		if isMigrated {
 			continue
 		}